Test: 워크스페이스 생성 key prefix 테스트
